backend/requirements.txt
fastapi
uvicorn
supabase
requests
python-dotenv
httpx
apscheduler
pydantic
python-jose[cryptography]
